requirements.txt
fastapi~=0.119.0
pydantic~=2.12.2
SQLAlchemy~=2.0.44
python-dotenv~=1.1.1
psycopg2-binary~=2.9.11
httpx~=0.28.1
uvicorn~=0.37.0